login, logout, register added

+++ week2/database/deleteData.php
@@ -1,32 +1,18 @@
 <?php
+session_start();
+if (!isset($_SESSION['user'])) {
+    header('Location: index.php');
+    exit;
+}
+
 global $DBH;
 require_once 'dbConnect.php';
 
 if(isset($_GET['id'])) {
     $DBH->beginTransaction();
     $data = [
-        'media_id' => $_GET['id']
+        'media_id' => $_GET['id'],
     ];
-    // delete file from server
-    $sql = 'SELECT filename FROM MediaItems WHERE media_id = :media_id';
-    try {
-        $STH = $DBH->prepare($sql);
-        $STH->execute($data);
-        $STH->setFetchMode(PDO::FETCH_ASSOC);
-        $row = $STH->fetch();
-        $filename = $row['filename'];
-        $destination = __DIR__ . '/uploads/' . $filename;
-        if (!unlink($destination)) {
-            echo "Could not delete file from the server.";
-            $DBH->rollBack();
-            exit;
-        }
-    } catch (PDOException $e) {
-        echo "Could not get filename from the database.";
-        file_put_contents('PDOErrors.txt', 'deleteData.php - ' . $e->getMessage(), FILE_APPEND);
-        $DBH->rollBack();
-        exit;
-    }
 
     // delete likes
     $sql = 'DELETE FROM Likes WHERE media_id = :media_id';
@@ -78,14 +64,41 @@ if(isset($_GET['id'])) {
         exit;
     }
 
+    $data['user_id'] = $_SESSION['user']['user_id'];
+
+    // delete file from server
+    $sql = 'SELECT filename FROM MediaItems WHERE media_id = :media_id AND user_id = :user_id';
+    try {
+        $STH = $DBH->prepare($sql);
+        $STH->execute($data);
+        $STH->setFetchMode(PDO::FETCH_ASSOC);
+        $row = $STH->fetch();
+        $filename = $row['filename'];
+        $destination = __DIR__ . '/uploads/' . $filename;
+        if (!unlink($destination)) {
+            $DBH->rollBack();
+            header('Location: home.php?success=File delete failed');
+            exit;
+        }
+    } catch (PDOException $e) {
+        echo "Could not get filename from the database.";
+        file_put_contents('PDOErrors.txt', 'deleteData.php - ' . $e->getMessage(), FILE_APPEND);
+        $DBH->rollBack();
+        exit;
+    }
+
     // delete MediaItems
-    $sql = 'DELETE FROM MediaItems WHERE media_id = :media_id';
+    $sql = 'DELETE FROM MediaItems WHERE media_id = :media_id AND user_id = :user_id';
 
     try {
         $STH = $DBH->prepare($sql);
         $STH->execute($data);
         $DBH->commit();
-        header('Location: index.php?success=Item deleted');
+        if($STH->rowCount() > 0) {
+            header('Location: home.php?success=Item deleted');
+        } else {
+            header('Location: home.php?success=Not your item.');
+        }
     } catch (PDOException $e) {
         echo "Could not delete data from the database.";
         file_put_contents('PDOErrors.txt', 'deleteData.php - ' . $e->getMessage(), FILE_APPEND);
@@ -93,5 +106,5 @@ if(isset($_GET['id'])) {
         exit;
     }
 } else {
-    header('Location: index.php?success=No hacking allowed.');
+    header('Location: home.php?success=No hacking allowed.');
 }
